+++ client/src/components/library/index.js
@@ -2,3 +2,5 @@ export { default as CategorySidebar } from './CategorySidebar';
 export { default as SearchAndSort } from './SearchAndSort';
 export { default as ActionButtons } from './ActionButtons';
 export { default as ComponentList } from './ComponentList';
+export { default as CadFieldSection } from './CadFieldSection';
+export { default as CadFilePickerModal } from './CadFilePickerModal';